moved favouritesContainer folder into favourites folder
