+++ component/Favourites/FavouritesContainer/FavouritesContainer.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import classes from './FavouritesContainer.module.css';
-import StateContext from '../../useContext/StateContext';
-import PreviewList from '../Recipes/Preview/PreviewList';
-import LoadingSpinner from '../ui/loadingSpinner/LoadingSpinner';
+import StateContext from '../../../useContext/StateContext';
+import PreviewList from '../../Recipes/Preview/PreviewList';
+import LoadingSpinner from '../../ui/loadingSpinner/LoadingSpinner';
 
 export default function FavouritesContainer() {
   const { favouritesList } = StateContext();
